chore: add changeset for schemars schema generation
