+++ crates/monochange_schema_gen/docs/src/schemas/monochange.schema.json
