+++ crates/monochange_schema_gen/docs/src/schemas/monochange.v0.1.schema.json
