+++ crates/monochange_schema_gen/docs/src/schemas/release-record.schema.json
