+++ crates/monochange_schema_gen/docs/src/schemas/release-record.v0.1.schema.json
